Alternative benchmarking script:
* Disabled the grid-param-related runs - they're mostly useless as long as we're I/O-bound (rather than checking compute time with nvprof)
* No longer skipping runs just because the plot.R file exists; we currently have the same file for multiple options, so that's broken anyway

+++ scripts/altbench.py
@@ -65,9 +65,6 @@ def run_test(filename_for_plot = None, raw_results_file = None, mean_results_fil
 	if not options: options = default_options
 	if not placement: placement = default_placement
 
-	if os.path.isfile(os.path.join(results_dir, filename_for_plot)):
-		return
-
 	syscall("""${BINARY} ${OPTIONS} --streams=${STREAMS} --scale-factor=${SF} --tuples-per-kernel-launch=${TUPLES} --tuples-per-thread=${VALUES} --threads-per-block=${THREADS} --hash-table-placement=${PLACEMENT}""".replace(
 		"${BINARY}", binary).replace(
 		"${OPTIONS}", options).replace(
@@ -81,7 +78,8 @@ def run_test(filename_for_plot = None, raw_results_file = None, mean_results_fil
 		return
 
 	full_plot_fn = os.path.join(results_dir, filename_for_plot)
-	copyfile('results.csv', full_plot_fn)
+	if not os.path.isfile(full_plot_fn):
+		copyfile('results.csv', full_plot_fn)
 
 	if raw_results_file:
 		compressed_data = (options.find("apply-compression") != -1)
@@ -120,14 +118,14 @@ for opt in options:
 	for p in placements:
 		run_test(filename_for_plot = os.path.join(options_basename,'%s.csv' % p), raw_results_file = raw_fn, mean_results_file = mean_fn, options = opt, placement = p)
 
-gp_basename='grid_params'
-raw_fn = os.path.join(results_dir, '%s_raw.csv' % gp_basename)
-mean_fn = os.path.join(results_dir, '%s.csv' % gp_basename)
-init_results_files(gp_basename)
-for vals in tuples_per_thread:
-	for threads in threads_per_block:
-		for p in placements:
-			if default_tuples_per_launch >= vals * threads * cores_per_gpu * min_keep_busy_factor :
-				run_test(raw_results_file = raw_fn, mean_results_file = mean_fn, filename_for_plot=os.path.join(gp_basename, 'vals_%s-threads_%s-placement_%s.csv' % (str(vals), str(threads), p)), vals = vals, threads = threads, placement = p)
-			else:
-				print ("%s tuples per thread and %s threads per block are too many for %s tuples per launch - there would not be enough blocks to keep the GPU busy" % (vals, threads, default_tuples_per_launch))
+#gp_basename='grid_params'
+#raw_fn = os.path.join(results_dir, '%s_raw.csv' % gp_basename)
+#mean_fn = os.path.join(results_dir, '%s.csv' % gp_basename)
+#init_results_files(gp_basename)
+#for vals in tuples_per_thread:
+#	for threads in threads_per_block:
+#		for p in placements:
+#			if default_tuples_per_launch >= vals * threads * cores_per_gpu * min_keep_busy_factor :
+#				run_test(raw_results_file = raw_fn, mean_results_file = mean_fn, filename_for_plot=os.path.join(gp_basename, 'vals_%s-threads_%s-placement_%s.csv' % (str(vals), str(threads), p)), vals = vals, threads = threads, placement = p)
+#			else:
+#				print ("%s tuples per thread and %s threads per block are too many for %s tuples per launch - there would not be enough blocks to keep the GPU busy" % (vals, threads, default_tuples_per_launch))
